requirements.txt
torch
torchvision
numpy
opencv-python
scikit-learn
Pillow
matplotlib
tqdm
dtw-python
filterpy
numba